requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.1.7
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.1
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
-e git+ssh://git@github.com/yoninazarathy/shalev.git@2e5a5bec29da0cd04213e8224a753e566514dd1d#egg=shalev
sniffio==1.3.1
structlog==25.4.0
termcolor==3.2.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yaspin==3.3.0
